+++ Changelog.md
@@ -16,6 +16,8 @@
 
 ### New Features
 
+- [#483]: Implement `read_text_into()` and `read_text_into_async()`.
+
 ### Bug Fixes
 
 - [#939]: Fix parsing error of the tag from buffered reader, when the first byte `<`
@@ -24,6 +26,7 @@
 
 ### Misc Changes
 
+[#483]: https://github.com/tafia/quick-xml/issues/483
 [#936]: https://github.com/tafia/quick-xml/pull/936
 [#939]: https://github.com/tafia/quick-xml/issues/939